python-api/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pika==1.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
boto3>=1.34.0
python-dotenv==1.1.1
requests==2.32.3

